requirements.txt
google-genai
requests
pydantic
python-dotenv
streamlit